Version: 1.10
Switched freeimage over to multithreaded dll at compile time and static linked it, avoiding needing the dll present to run the app.
Also cleared all but 3 level 1 compiler warnings.
Also updated all prerequisite libraries to latest version.

+++ src/Modules/Graphics/Resources/Effects/SSAO.h
@@ -42,7 +42,7 @@ public:
 		m_shapeQuad->addCallback(this, [&]() mutable {
 			m_quadVAOLoaded = true;
 			m_shapeQuad->updateVAO(m_quadVAO);
-			const GLuint quadData[4] = { m_shapeQuad->getSize(), 1, 0, 0 }; // count, primCount, first, reserved
+			const GLuint quadData[4] = { (GLuint)m_shapeQuad->getSize(), 1, 0, 0 }; // count, primCount, first, reserved
 			m_quadIndirectBuffer.write(0, sizeof(GLuint) * 4, quadData);
 		});
